requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pinecone-client>=3.0.0,<4.0.0
sentence-transformers==2.2.2
pydantic==2.4.2
requests>=2.28.0
pypdf==3.17.0
python-multipart==0.0.6
langchain==0.0.335
langchain_community==0.0.13
langchain_pinecone==0.0.1
httpx==0.25.0
google-generativeai>=0.3.0
